+++ src/components/RestaurantList.js
@@ -1,0 +1,27 @@
+import createElement from '../utils/createElement.js';
+import createRestaurantItem from './RestaurantItem.js';
+
+function createRestaurantList(data) {
+  const restaurantList = createElement('ul', 'restaurant-list');
+
+  const fragment = new DocumentFragment();
+  data.forEach((data) => {
+    const restaurantItem = createRestaurantItem(data);
+    fragment.appendChild(restaurantItem);
+  });
+
+  restaurantList.appendChild(fragment);
+
+  return restaurantList;
+}
+
+function updateRestaurantList(inputData) {
+  const $restaurantList = document.querySelector('.restaurant-list');
+  const $restaurantItem = createRestaurantItem(inputData);
+
+  $restaurantList.appendChild($restaurantItem);
+
+  return $restaurantList;
+}
+
+export { createRestaurantList, updateRestaurantList };
